refactored player board and computer board components
